# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OSAM Voice Banking API
  description: |
    Voice-enabled banking API for IPPB integration.
    Supports Hindi, Tamil, and English voice commands.
  version: 1.0.0
  contact:
    name: RBI Innovation Hub
    url: https://rbihub.in

servers:
  - url: http://localhost:8002
    description: Local development
  - url: https://osam-voice-banking.onrender.com
    description: Production (Render)

paths:
  /api/health:
    get:
      summary: Health check
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
                  service: { type: string, example: OSAM Voice Banking }
                  version: { type: string, example: 1.0.0 }

  /api/languages:
    get:
      summary: List supported languages
      operationId: listLanguages
      responses:
        '200':
          description: Languages list
          content:
            application/json:
              schema:
                type: object
                properties:
                  languages:
                    type: array
                    items:
                      type: object
                      properties:
                        code: { type: string, example: hi }
                        name: { type: string, example: हिन्दी }

  /api/voice/process:
    post:
      summary: Process voice command
      operationId: processVoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text: { type: string, description: Transcribed text or typed input }
                audio: { type: string, description: Base64 encoded audio (optional) }
                language: { type: string, enum: [hi, ta, en], default: hi }
                session_id: { type: string, description: Session for multi-turn conversations }
      responses:
        '200':
          description: Processed voice command
          content:
            application/json:
              schema:
                type: object
                properties:
                  intent: { type: string, enum: [balance, loan, branch, help, goodbye, unknown] }
                  transcript: { type: string }
                  response_text: { type: string }
                  requires_input: { type: boolean }
                  session_id: { type: string }
                  audio_url: { type: string, nullable: true }

  /api/balance:
    post:
      summary: Check account balance
      operationId: checkBalance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                account_number: { type: string, example: "123456" }
                language: { type: string, enum: [hi, ta, en], default: hi }
      responses:
        '200':
          description: Balance information
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance: { type: number, nullable: true }
                  response_text: { type: string }
                  audio_url: { type: string, nullable: true }

  /api/loan/apply:
    post:
      summary: Apply for a loan
      operationId: applyLoan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: number, example: 50000 }
                purpose: { type: string, example: business }
                language: { type: string, enum: [hi, ta, en], default: hi }
      responses:
        '200':
          description: Loan application result
          content:
            application/json:
              schema:
                type: object
                properties:
                  application_id: { type: string, example: OSAM-A1B2C3D4 }
                  status: { type: string, example: submitted }
                  response_text: { type: string }
                  audio_url: { type: string, nullable: true }

  /api/branch/nearby:
    post:
      summary: Find nearby branch
      operationId: findBranch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                location: { type: string, example: delhi }
                language: { type: string, enum: [hi, ta, en], default: hi }
      responses:
        '200':
          description: Nearby branch information
          content:
            application/json:
              schema:
                type: object
                properties:
                  branches:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        address: { type: string }
                        distance: { type: number }
                  response_text: { type: string }
                  audio_url: { type: string, nullable: true }

  /api/audio/{audio_id}:
    get:
      summary: Get generated audio file
      operationId: getAudio
      parameters:
        - name: audio_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Audio file (MP3)
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary

  /api/compliance/rules:
    get:
      summary: Get IS 17802 compliance rules
      operationId: complianceRules
      responses:
        '200':
          description: List of accessibility compliance rules

  /api/compliance/check:
    get:
      summary: Run accessibility compliance check
      operationId: complianceCheck
      responses:
        '200':
          description: Accessibility compliance report

  /api/voice/upload:
    post:
      summary: Upload audio file for transcription
      operationId: uploadAudio
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                language:
                  type: string
                  enum: [hi, ta, en]
                  default: hi
      responses:
        '200':
          description: Transcription result
          content:
            application/json:
              schema:
                type: object
                properties:
                  filename: { type: string }
                  transcript: { type: string }
                  intent: { type: string }
                  language: { type: string }